Moved storage addresses from mod into storage.
commit-id:0537c4c1
